+++ src/Hacks/ShowLayout.cpp
@@ -45,16 +45,20 @@ class $modify(GameObject) {
         }
     }
 };*/
-/*
 
 #include <Geode/Geode.hpp>
 #include <Geode/modify/PlayLayer.hpp>
+#include <Geode/modify/GJBaseGameLayer.hpp>
 #include <Geode/modify/GameObject.hpp>
 #include <Geode/modify/EffectGameObject.hpp>
+#include "../Client/Client.h"
 
 using namespace geode::prelude;
 
-std::vector<char const*> objs = {
+Module* showLayout = nullptr;
+Module* showLayoutCamera = nullptr;
+
+std::vector<std::string> objs = {
 
     // SHADERS
     "edit_eShaderBtn_001.png",
@@ -76,7 +80,17 @@ std::vector<char const*> objs = {
     "edit_eSh_EditColorBtn_001.png",
     "edit_eSh_SplitScreenBtn_001.png",
 
-    // CAMERA
+    // Alpha Trigger
+    "edit_eAlphaBtn_001.png",
+
+    // Colour Trigger
+    "edit_eTintCol01Btn_001.png",
+
+    // Gradient
+    "edit_eGradientBtn_001.png",
+};
+
+std::vector<std::string> camera = {
     "edit_eCamModeBtn_001.png",
     "edit_eGPOffsetBtn_001.png",
     "edit_eEdgeBtn_001.png",
@@ -85,12 +99,96 @@ std::vector<char const*> objs = {
     "edit_eOffsetBtn_001.png",
     "edit_eStaticBtn_001.png",
     "edit_eZoomBtn_001.png",
+};
 
-    // idk what to call this
-    "edit_eAlphaBtn_001.png",
+class $modify (PlayLayer)
+{
+    CCSprite* background;
+    GJGroundLayer* ground1;
+    GJGroundLayer* ground2;
 
-    // AREA
-    "edit_eTintCol01Btn_001.png",
+    virtual TodoReturn postUpdate(float dt)
+    {
+        PlayLayer::postUpdate(dt);
+
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!showLayout->enabled)
+            return;
+
+        if (!m_fields->background)
+            m_fields->background = getChildOfType<CCSprite>(getChildOfType<CCNode>(this, 0), 0);
+
+        if (!m_fields->ground1)
+            m_fields->ground1 = getChildOfType<GJGroundLayer>(getChildOfType<CCNode>(this, 0), 0);
+
+        if (!m_fields->ground2)
+            m_fields->ground2 = getChildOfType<GJGroundLayer>(getChildOfType<CCNode>(this, 0), 0);
+
+        if (m_fields->background && m_fields->ground1 && m_fields->ground2)
+        {
+            m_fields->background->setColor(ccc3(40, 125, 255));
+
+            m_fields->ground1->updateGround01Color(ccc3(0, 102, 255));
+            m_fields->ground1->updateGround02Color(ccc3(0, 102, 255));
+
+            m_fields->ground2->updateGround01Color(ccc3(0, 102, 255));
+            m_fields->ground2->updateGround02Color(ccc3(0, 102, 255));
+        }
+    }
+};
+
+class $modify (GJBaseGameLayer)
+{
+    void createMiddleground(int p0)
+    {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            GJBaseGameLayer::createMiddleground(p0);
+    }
+
+    void createBackground(int p0)
+    {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        GJBaseGameLayer::createBackground(PlayLayer::get() && showLayout->enabled ? 0 : p0);
+    }
+
+    void createGroundLayer(int p0, int p1)
+    {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        GJBaseGameLayer::createGroundLayer(PlayLayer::get() && showLayout->enabled ? 0 : p0, PlayLayer::get() && showLayout->enabled ? 0 : p1);
+    }
+
+    virtual TodoReturn updateColor(cocos2d::ccColor3B& p0, float p1, int p2, bool p3, float p4, cocos2d::ccHSVValue& p5, int p6, bool p7, EffectGameObject* p8, int p9, int p10)
+    {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            GJBaseGameLayer::updateColor(p0, p1, p2, p3, p4, p5, p6, p7, p8, p9, p10);
+    }
 };
 
 class $modify(EffectGameObject) {
@@ -98,8 +196,23 @@ class $modify(EffectGameObject) {
 		if (!EffectGameObject::init(p0)) 
             return false;
 
-        if (std::find(objs.begin(), objs.end(), p0) != objs.end())
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            return true;
+
+        if (std::find(objs.begin(), objs.end(), std::string(p0)) != objs.end())
             return false;
+
+        if (!showLayoutCamera->enabled)
+        {
+            if (std::find(camera.begin(), camera.end(), std::string(p0)) != camera.end())
+                return false;
+        }
 
 		return true;
 	}
@@ -107,6 +220,15 @@ class $modify(EffectGameObject) {
 
 class $modify(GameObject) {
     void setVisible(bool v) {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            return GameObject::setVisible(v);
+        
         //m_hasGroupParent == 0
         std::vector<int> outerPortal = {};
         //i really dont want to have to check every single object id
@@ -116,4 +238,41 @@ class $modify(GameObject) {
             GameObject::setVisible(v);
         }
     }
-};*/
+
+    void setOpacity(unsigned char g) {
+
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+		GameObject::setOpacity(g);
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            return;
+
+		if (true && m_objectType != GameObjectType::Decoration) { // true being layout enabled
+			CCSpritePlus::setOpacity(255);
+			if (m_glowSprite)
+				m_glowSprite->setOpacity(255);
+		}
+	}
+
+	void setObjectColor(const cocos2d::ccColor3B& cor) {
+        if (!showLayout)
+        {
+            showLayout = Client::GetModule("show-layout");
+            showLayoutCamera = showLayout->options[0];
+        }
+
+        if (!PlayLayer::get() || !showLayout->enabled)
+            return GameObject::setObjectColor(cor);
+
+		if (true && m_objectType != GameObjectType::Decoration) {
+			GameObject::setObjectColor(ccc3(255, 255, 255));
+		} else {
+			GameObject::setObjectColor(cor);
+		}
+	}
+};